Changes on server: * Extracted login from index.ts file to handlers * Created connection handler * Created handlers folder and moved handlers
